The block action now sets flags for the engine and other rules to consume.
When an advisory block is triggered, TX.BLOCK is set to 1. Also
the tx.flags value has the IB_TX_BLOCK_ADVISE flag set.

When an immediate block is triggered the tx.flags IB_TX_BLOCK_IMMEDIATE
flag is set.

When a phase block is triggered the tx.flags IB_TX_BLOCK_PHASE flag is
set.

+++ engine/config-parser.c
@@ -489,7 +489,7 @@ _resume:
 			else if ( (*( fsm.p)) > *_mid )
 				_lower = _mid + 1;
 			else {
-				_trans += (unsigned int)(_mid - _keys);
+				_trans += (_mid - _keys);
 				goto _match;
 			}
 		}
@@ -512,7 +512,7 @@ _resume:
 			else if ( (*( fsm.p)) > _mid[1] )
 				_lower = _mid + 2;
 			else {
-				_trans += (unsigned int)((_mid - _keys)>>1);
+				_trans += ((_mid - _keys)>>1);
 				goto _match;
 			}
 		}
